raw-l2-rcv: Fix signal handling
Replace the simpler signal() function with sigaction(), because signal()
makes all system calls by default re-entrable upon a signal, which is
exactly what we don't want.

We want a system call (the recvmsg, to be precise) to exit immediately
with -EINTR and not have the program wait until it gets more data on the
socket before it closes itself.

This way we can kill the program with start-stop-daemon from
time-test.sh.

Fixes: 17c848f043f3 ("raw-l2-rcv: Avoid truncation on last output line when killed")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>