8021cb: Add some checks for system configuration
I had initially tested the 802.1CB setup on OpenIL, which is quite a
simple rootfs.
With systemd-based distributions (LS1028A BSP, OpenIL with Ubuntu) one
must configure system programs to not send anything on the VLAN
sub-interfaces, because we deliberately allow loops in VID 100 and just
use that for 802.1CB traffic.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>