8021cb: Print when configuration is successful
Currently tsntool prints a lot of junk even in non-interactive mode. I
had patched it so it no longer does, but it's going to take a while
until users get these patches, and it's confusing to see messages such
as:

  tsntool: invalid option -- '-'
  tsntool: invalid option -- 'd'
  echo reply:test netlink
  echo reply:1
  echo reply:swp1
  echo reply:0

which don't really mean anything.

I'm not going to redirect its output to /dev/null, because who knows
what real error I might be hiding. Just print this message and the user
will have more confidence that the tsntool commands worked.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>