time-test: Give board1 time to become PTP master
Because we only check port state once, the check_sync function may enter
the while loop waiting for the PTP clock to become synchronized, which
is never going to happen. So check port state in each iteration (the
master is going to be in LISTENING first until the BMCA finishes).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>