time-test: Don't hardcode the UTC offset
Instead retrieve it from the PTP management client.
Actually linuxptp currently doesn't even claim that the UTC offset is
valid (which it isn't... it's supposed to be 37 and not 36). However we
don't care, as long as we can translate from CLOCK_TAI to CLOCK_REALTIME
and back using this factor.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>