time-test: Add compensation arguments for Qbv on ENETC
These parameters are no-ops so far, since the memory writes simply
reflect the hardware defaults.  Will replace them once a clear
compensation formula is established.

What I'm trying to understand and reduce is the ENETC latency between a
Qbv gate open event and the corresponding frame's TX timestamp:

  Mean MAC TX latency (MAC TX - gate event time): 411 ns
  Standard deviation: 57 ns

So far it has been established that it is frame length and link speed
dependent.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>