time-test: Configure the Reference Manual values for PESMATOR
This effectively shifts the entire Qbv schedule by the specified amount
in nanoseconds, in order to compensate for MAC delays at various data
rates.

In the (MAC gate event - TX timestamp) diff there is still a ~30 ns
latency that is unaccounted for (considering that the TX timestamp is
taken after 7 preamble octets and 1 SFD).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>