time-test: Make sure Felix kernel patch for broadcast is applied
arping gets confused in the BSP 0.3 kernel, because all Felix port
drivers will respond to the broadcast ARP message sent to discover the
MAC address of eno2.

The application gets further confused, as it is expecting a single MAC
address to come from arping.

  "Could not read MAC address: File too large"

So just make sure this patch (taken from the internal pre-release git
tree) is applied.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>