time-test: Make Felix and ENETC synchronize over Ethernet
The two PTP clocks are required to be in sync for some applications,
e.g. to create a PTP boundary clock that spans both ENETC and Felix
ports. Another (perhaps less honourable, but more relevant to these
scripts) reason is that we are currently using eno2 to inject traffic
into the switch's Qbv queues, instead of the more idiomatic swpN way
which requires QoS support through the injection/extraction header
(using VLAN sub-interfaces on eno2 achieves this by transmitting in-band
QoS hints through the VLAN PCP). Remember we are polling the PTP clock
in raw-l2-send, so if we're sending over eno2, it still needs to be in
sync with the Felix clock which governs the Qbv schedule.

Anyway, software sync performance (through phc2sys) is not as good as
can be. Unfortunately it looks like LS1028A has no dedicated hardware
(i.e. PPS connection) between ENETC and Felix so they can be kept in
sync.

Fortunately Richard Cochran suggested something that works as a perfect
workaround: use the untagging eno2-swp4 port pair as a loopback
connection for running 2 ptp4l instances. Since these use hardware
timestamping, the sync offset is much lower.

https://www.spinics.net/lists/netdev/msg594535.html

Add two ptp4l configuration files that are largely stolen from the
automotive master and slave profiles. There is no BMCA needed, it is a
completely isolated process from the ptp4l running on other ports. We
also need separate UNIX domain sockets for these to not interact with
pmc trying to get data from the main ptp4l instance.

Suggested-by: Richard Cochran <richardcochran@gmail.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>