raw-l2-send: raw-l2-rcv: Prepare for Y2038 timespec wraparound
This patch reduces the usage of "struct timespec" in the application to
a minimum (basically strictly to the interaction with glibc). The time_t
tv_sec is long int, therefore 32-bit on 32-bit systems, which will wrap
around in 2039.

The glibc library is in the process of conversion:
https://sourceware.org/glibc/wiki/Y2038ProofnessDesign

This further simplifies the handling of time in the program somewhat,
and uses a similar API as kernel drivers, which deal with u64/s64
directly.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>