raw-l2-send: raw-l2-rcv: Add a software timestamp
The awk parser is currently calculating the OS TX latency as (OS TX -
scheduled TX time), where OS TX is the '[%s] Sent frame' portion of the
raw-l2-send output (more-or-less the time of enqueue into the
application socket) and the scheduled TX time is the 'scheduled for %s'
portion of the raw-l2-send output (which is the UTC translation of the
MAC gate event).

But this is not really what the OS TX latency is. In fact there are 2
parts to it:
- Application latency: wakeup time to socket enqueue time. This is not
  printed right now.
- Driver latency: socket enqueue time to driver software timestamp. This
  is not printed either.

And this changes the meaning of the other variables too:
- MAC TX latency can be (re-)defined as (HW timestamp - SW timestamp -
  UTC offset).
- The old meaning of MAC TX latency can be now called 'MAC TX accuracy',
  the time it takes the MAC to put the frame on the wire after the Qbv
  gate event has fired.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>