raw-l2-send: Add a shift-time argument
Previously it was only possible to send cyclic traffic at the beginning
of a schedule. Otherwise, the user needed to modify the base-time (skew
it compared to the MAC schedule, to obtain a permutation where its own
time slot is the first in the schedule).

Now the user is able to specify the same base-time for raw-l2-send as
for the MAC schedule, and specify an extra --shift-time which represents
the cumulative offset of raw-l2-send's time slot compared to the
beginning of the cycle.

So now, raw-l2-send is scheduled for the Nth cycle at the following
PTP time:

base-time + shift-time + (N * cycle-time) - advance-time.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>