time-test: Rename system services for Buildroot compatibility
When running on Buildroot, we certainly don't want to have two sets of
linuxptp services running at the same time.

Moreover, when running the demo scripts between two boards, one running
OpenIL, one running Buildroot, the ssh from board A to B will try to
check services that are named differently. We certainly don't want to
over-complicate the scripts and check the distribution.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>