raw-l2-send: Fix SCHED_DEADLINE scheduler attributes
Remove a previous workaround which makes sched_setattr not fail even if
the user does not specify any --advance-time. That means the
advance-time is implicitly zero, which is fundamentally wrong: there is
no time for raw-l2-send to do its work, because it is already too late.
Passing 0 to the sched_deadline would make the kernel validator return
-EINVAL. So the workaround was to reverse the meaning of the advance
time when specifying to the scheduler.

The correct fix is to change the default advance-time value, from 0 to
the length of the cycle itself. That has been done, so we can remove the
workaround.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>