raw-l2: Make the processing of TX timestamps asynchronous
It was seen with kernelshark that sleeping while waiting to receive the
TX timestamps could consume even 20 us.

So right now we are polling the error queue of the socket with a timeout
of zero from the hotpath (just to make sure that if a timestamp _is_
available, we can consume it without sleeping). We do that to make sure
that the socket buffer doesn't overflow.

The TX timestamps that are not acknowledged during the hotpath are
acknowledged at the end.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>