raw-l2: Fix the definition and usage of future_base_time
The implementation from sja1105_tas.c has been dropped in favor of
sch_taprio.c. The key is to use div64_s64, to avoid an integer overflow
that is currently occuring while dividing to the cycle_time.

This patch also draws some math functions from the Linux kernel as
dependencies.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>