raw-l2-common: Fix printing of timestamps
The timestamps are kept internally as __s64 (signed long long int).
Because I wanted to play nice and print negative time deltas in a pretty
format, I used the abs() function thinking it is a macro.

But man 3 abs shows it is a stdlib.h function that takes an int as
argument, so overflow is occurring when feeding s64 to it. This caused
all sorts of strange behavior, such as time that appears to be going
backwards (and jumping at times).

Use the proper abs() function that takes long long ints as argument
(llabs).

Fixes: 939e2f2a94d8 ("raw-l2: Use native fixed-size integer types exported by the Linux kernel UAPI")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>