8021cb: Use raw write to /proc/sys/net instead of sysctl
The sysctl tool blindly replaces . with / when computing the sysfs path
to write to. But our VLAN sub-interfaces have dots in their name, and
this would result in:

sysctl: cannot stat /proc/sys/net/ipv6/conf/eno2/101/disable_ipv6: No such file or directory

For some reason, this error only manifests itself on LSDK, but not on
the OpenIL rootfs.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>