raw-l2: Make timestamping optional
This patch introduces a new boolean option: -T|--no-ts, which specifies
that we are not interested in timestamps taken from the kernel (neither
sw nor hw) just in sending and receiving packets.

This is needed in order to work around the fact that we are using
raw-l2-send and raw-l2-rcv for the 802.1CB demo as well, but the LSDK
19.09 kernel does not ship CONFIG_FSL_ENETC_HW_TIMESTAMPING enabled by
default. And because it is not mandatory, it is better that we make it
optional at the application layer too.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>