8021cb: Adapt to a 2-board setup
The 3-board setup currently has some limitations.

With 3 boards (A, B, C) in a ring and 2 streams (S: A -> B, T: B -> A)
that are actually redundant, they look like this:
A -> B: member S1: A->B and member S2: A->C->B
B -> A: member T1: B->A and member T2: B->C->A

The trouble, of course, is with the poor switch in the middle (C).
By and large, switches work on the assumption that the port where the
destination host is is the same as the port from which some traffic
coming from that host was received in the past. Or, in short, they do
"L2 address learning".

So board C receives a packet from member stream S2. It says "ok, so host
A is connected to the port on my left".
Then it receives a packet from member stream T2, towards host A. It
knows that the route towards host A is the port on its left. "Wait a
minute, you want me to send this packet back on the port I just got it
from (aka do hairpinning)? Sorry, no." and it drops it.

Basically, any intermediary switch for a bidirectional 802.1CB stream
will drop traffic in one of the directions, due to its hardware address
learning process.

So the workaround I've tried so far is to have a VLAN per destination.
This way, stream S2 will be directed towards B so it will carry the VID
corresponding to host B, and stream T2 will carry the VID corresponding
to host A.
Because the switch supports Independent VLAN Learning, the switch will
learn the destination towards host A in VLAN B only, so it won't drop
packets to A when they are tagged with VLAN A.

This workaround accomplishes its purpose (the reverse 802.1CB stream is
no longer dropped by the intermediary switch) but it makes bidirectional
communication a lot more difficult for upper-layer applications, since
the VID needs to have different values for the 2 directions of
communication.

So a better workaround might be to just disable address learning in all
switch ports:

bridge link set dev swp0 learning off

The trouble is that this isn't yet implemented in the Felix kernel
driver.

So, for the purposes of demonstrating bidirectional traffic, it might be
easier to just create a 2-board ring, where this problem doesn't exist
because there is no intermediary switch.

Signed-off-by: Bob Cutler <bob.cutler@hpe.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>