isochron: send PTP packets instead of custom ones
DSA is more strict about the frames it allows to be timestamped, and
runs a BFP filter (ptp_classify_raw) to determine them.

To be compatible with as much hardware as possible, construct PTP event
(sync) frames to be timestamped, and general (follow-up) frames for
testing various hardware at large (not necessary for the demo scripts).

The assumption is that if the hardware supports any sort of timestamping
at all, it will certainly be able to timestamp PTP frames.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>