8021cb: ignore kernel errors when trying to clear stream table
The logic from the clear_stream_table() function is to make sure that we
are deleting everything that was configured before, so these scripts can
run in a clean slate environment without rebooting the board.

But a kernel-side change has been introduced recently, and the first
time the tsntool cbstreamidset --disable command is run, it fails:

tsntool> cbstreamidset --index 0 --nullstreamid --streamhandle 0 --device swp0 --disable

echo reply:swp0
echo reply:-22

This error corresponds to this kernel-side piece of code:

static int felix_cb_streamid_set(struct net_device *ndev, u32 index, bool enable,
				 struct tsn_cb_streamid *streamid)
{
	...

	if (!enable) {
		stream = stream_table_get(index);
		if (!stream)
			return -EINVAL;

So it will error out if the stream doesn't exist (and it doesn't,
because we've just booted the board).

So ignore errors and continue running the script.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>