8021cb: delete the patches that were integrated into OpenIL
As of commit 315b5e616040 ("Merge pull request #165 in
DNIND/industry-linux from
~NXA24537/industry-linux:bugfix/INDLINUX-1634-felix-802.1cb-stream-splitting-broken-when-2-ssids-are-used
to linux-5.4.y"), patches 2 to 7 have been merged into the OpenIL
kernel. So now we only have the DTS conversion, and a related patch
(port flushing) which is needed only with the DTS conversion (otherwise
flow control is not enabled).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>