isochron: send: log time to trace marker
This modifies the previously simplistic function to write to the trace
buffer, so that it includes the time at which the write was made.

This is useful because ftrace data is collected in the CLOCK_MONOTONIC
time domain (which on Linux counts time since boot, minus time spent
suspended), but the application will be synchronized to the network TAI
time. So it is more useful to know what the TAI time was when a certain
event happened.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>