isochron: record wakeup latency
One of the major sources of latency, even with PREEMPT_RT, is the
scheduler (due to cache misses, preemption disabled, IRQs disabled).

When clock_nanosleep() expires at the requested advance time, its
kernel-side hrtimer raises an IRQ and that ends up scheduling the
application. But if we measure the system clock in the application right
after clock_nanosleep() expired, we'll see that it is 10 us, maybe 20 us,
larger than the requested expiry time. In fact this is also the strategy
used by cyclictest to measure the system's responsiveness.

Add a metric in this application for the wake-up time, so that the user
knows where to look in case that's what's going on.

As for remedies? Boost priority, change scheduling policy, isolate CPU,
run trace-cmd -p irqsoff to find kernel bugs, etc etc.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>