isochron: back to sending custom packets
Commit 3a1187344e28 ("isochron: send PTP packets instead of custom ones")
didn't age so well. In practice we aren't fooling anyone and not gaining
anything by sending PTP packets.

Yes, it's true that if we send custom packets and not normal PTPs then
we lose compatibility with DSA switches, but on the other hand, one of
the big use cases for the isochron program was to measure the latency of
said switches, i.e. treating them as devices under test rather than
testing equipment. And most DSA switches, when they see a PTP packet,
they'll trap it to the CPU instead of forwarding it. So no testing of
latency is possible with a normal PTP packet.

And now, we are running into the situation that we also need to encode
the sender's wakeup time into the transmitted packet (we are doing this
because it's the cheapest way to store that data in the hot path).
There's no field in the PTP header that we can hack for that purpose. So
it's about time that we return to a custom header for our application.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>