isochron: hide the statistics that have to do with software timestamping
At least on enetc, there seem to be some unexplained issues with
software timestamping, and the values are not very reliable / useful
when used in this application.

Note that we are still taking these timestamps, just not showing them.
One reason for taking them is to figure out whether CLOCK_TAI is in sync
with the PHC or not (the software timestamps are still accurate enough
to figure that out, just not more than that).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>