isochron: transmit the log buffer size when collecting statistics
This is a more wide-ranging change, because we need to be more careful
about the size of the log buffer. Currently it can happen that the
static allocation is not enough, for large tests. So we need to make the
buffer size configurable, depending on the size of the test.

The first 4 bytes of data transmitted on the stats fd are now the size
of the buffer (before, that was implicit based on when the server would
close the TCP connection).

We are also adding an argument to the receiver, which says how many
packets to expect per test. This indicates how large the buffer needs to
be.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>