isochron: boost priority temporarily within the application
At the time of commit 540af3711ffa ("raw-l2-send: Use chrt instead of
sched_setattr"), isochron (by then raw-l2-send) was a much simpler
program. In the meantime it has gained support for collecting statistics
from the receiver automatically, which has obsoleted much of the
time-test.sh script.

The problem with setting the priority externally, as we are doing right
now with chrt, is that the statistics collection process also takes
place at this high priority. And if we have a lot of statistics to
collect (100k packets), we're going to be running a CPU intensive task
as the highest priority task in the system. If that takes more than 21
seconds, it can easily create RCU stalls in the kernel.

We only need to run with a high scheduling priority during the hot path,
not when collecting and dissecting statistics. So it makes sense to
boost the priority by ourselves, and restore it afterwards.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>