isochron: send: drain potentially old packets from the isochron receiver
When a previous sender exited without collecting statistics, they are
still buffered on the receiver, which will attempt to send them next
time a sender connects, even if those test packets did not belong to it.

Create a dummy statistics connection before initiating the test, to make
sure old logged packets are discarded.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>