isochron: use CLOCK_TAI everywhere
Instead of hardcoding the UTC-TAI offset to 37 seconds, let's rely on
the kernel's TAI offset. In case the system does not have an NTP daemon
that has set that offset, also introduce a -O option for convenience
which sets the system offset to the required value.

Note that software packet timestamps are hardcoded in CLOCK_REALTIME, so
we still need to convert those.

Code is loosely based upon tools/testing/selftests/timers/set-tai.c.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>