isochron: add a script for detecting presence of SOF_TIMESTAMPING_OPT_TX_SWHW
Since SOF_TIMESTAMPING_OPT_TX_SWHW is defined as an enum in linux/net_tstamp.h,
our #ifndef check doesn't quite do the trick, since that only works for
preprocessor-defined macros.

In order to detect whether the linux/net_tstamp.h provided by our
toolchain (${CC}, ${CFLAGS}) provides that enum value or not, we must
write a small program that tries to access that enum, which is compiled
with the same toolchain as isochron itself. If the program fails,
isochron will fail too. If we succeed to compile it, we define a
HAVE_TX_SWHW macro which we pass through CFLAGS into isochron. Then we
know whether we need to define the missing shim for
SOF_TIMESTAMPING_OPT_TX_SWHW or not.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>