isochron: Make signal_received static.
The variable signal_received is globally defined in rcv.c and send.c and
both objects are linked into the isochron binary. The linking fails
because the variable is declared twice and collides with the other one.

The variable is only used locally (within the compile unit) for the
purporse of signaling ctrl-c from the signal handler.

Make the variable static so it isn't exported outside of the compile
unit.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>