isochron: make isochron_print_stats() callable from outside send.c
The new stats.c code will have a struct prog_data definition too, but it
will be binary incompatible with the one from send.c, because that's
just how we roll.

So we need to remove any occurrence of struct prog_data from
isochron_print_stats and the functions it calls, and pass everything up
to the caller.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>