isochron: send: use IPv6 header definition from glibc
It looks like in old glibc versions (tested with gcc 5.4 and glibc 2.21),
the kernel headers and the libc headers were not compatible with one
another, resulting in a duplicate definition of struct in6_addr when
both were included in the same program:

"Re: Redefinition of struct in6_addr in <netinet/in.h> and <linux/in6.h>"
https://sourceware.org/legacy-ml/libc-alpha/2013-01/msg00694.html

This was fixed in later glibc versions, but with 2.21 it is still
broken, and isochron consequently fails to build with that redefinition
error.

To maintain compatibility with that libc version, we can simply drop the
kernel-provided linux/in6.h and use the IPv6 header from netinet/in.h.

Fixes: 004c09ddbab4 ("isochron: send UDP packets too")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>