isochron: fix sendmsg error path
When sendmsg fails with SO_TXTIME, we will call sk_receive on the data
socket's error queue in order to retrieve a relevant kernel message
about our misconfiguration.

All is well, except the kernel also provides us timestamps in the cmsg,
but we aren't prepared to handle them: we've passed a NULL pointer for
the timestamp structure.

So guard against accessing a NULL timestamp pointer in sk_receive.

Fixes: c0c0924b4358 ("isochron: add support for txtime")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>