isochron: send: fix warnings in trace()
gcc is complaining that we might write more than 24 bytes of the now_buf
into the prog->tracebuf. We have to tell it to not write more, and we
use the .precision printf format specifier for that.

Secondly, the snprintf will terminate the string in a NULL character,
but we need it to continue with a space. Before, we were printing an
extra space which would make snprintf overrun the buffer in a very
controlled manner. As gcc does warn about that, just stop the apparent
off-by-one and manually replace the trailing NULL byte of the time
string with a space, to let the line flow into the actual trace message.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>