isochron: send: explicitly convert the sync offsets and path delay to long long
We just want to check whether the mean path delay and mean sync offsets
are in the ballpark. If higher than one second, it's likely that they're
out of sync.

We use llabs to avoid checking both negative and positive, however the
average values are floating points with double precision. So just cast
them to long long when making the comparison.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>