isochron: send an u32 over the stats socket for versioning
[ yes, this breaks compatibility with isochron binaries up till now, but ]

It would be nice to know if the isochron sender and receiver binary have
the same understanding of the log buffer used for stats, as well as of
the isochron packet header.

To be absolutely clear that there are version incompatibilities, I'll be
maintaining a constant compiled in both programs which will be bumped
every time an incompatible change is made. The receiver sends this to
the sender and that compares it with what is compiled in it.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>