isochron: send: add initial ptpmon support
This is an initial implementation of a new feature called ptpmon, which
is a bare-bones PTP management client (similar to pmc).

For background, I had initially attempted to solve the synchronization
topic outside of isochron, first through heavy-duty scripting (text
scraping through ptp4l's and phc2sys's logs), then through a small
program called "checksync" which I tried to send to linuxptp:

"[Linuxptp-devel] [RFC PATCH 00/10] checksync: new program to wait for a
local or remote clock to synchronize"
https://www.mail-archive.com/linuxptp-devel@lists.sourceforge.net/msg05169.html
https://sourceforge.net/p/linuxptp/mailman/message/37366206/

The log scraping worked the best, in the sense that it offered a similar
solution for ptp4l and for phc2sys, but was ridiculously unportable.
The checksync program got some interesting negative feedback, the main
recommendations being

(a) to modify phc2sys so that it tells the kernel via adjtimex and
    struct timex :: maxerror / esterror the current offset between the
    CLOCK_REALTIME and the slave PHC. Then to extend the kernel UAPI to
    be able to call something similar to adjtimex with maxerror /
    esterror on a PHC.

(b) to use libpmc and embed the checksync logic into my own program

Feedback item (a) actually contains two distinct points, and I'm
actually tempted to say "nope" to both. To the phc2sys && esterror idea
because from a quick glance it looks like the esterror resolution is
microseconds which is 3 orders of magnitude less than isochron needs for
predictable results. And also because NTP daemons also mess around with
the esterror and maxerror. If I just look at that piece of information,
I wouldn't have any idea whatsoever as to what I'm synchronized to.
And to the PHC && esterror idea, actually the same issue applies there
too, the isochron receiver can have its PHC synchronized, but to some
other random PTP grandmaster than the isochron receiver, which would be
utterly useless to the actual goal of making sure they're in sync to one
another.

Feedback item (b) actually makes a good point in that there would be
certain benefits in integrating PTP management logic into isochron
itself, one would be that scripts could become much more self-contained
and setups hopefully way easier to reproduce. Also, the ability to check
for things that are much more fine-grained than what a person can
reasonably do in bash: such as monitor port states, ensure the isochron
sender and receiver are synced to the same GM, etc.

But on the other hand, the idea of linking to Erez Geva's libpmc is
comical, since it offers no C API my two options basically are:

(1) compile isochron with g++: ain't gonna fly
(2) write the C wrappers for libpmc myself and put them in isochron, aka
    the one thing that Erez did not want to do

This, plus the fact that linking isochron to an external library would
make compilation significantly more difficult, and the result is that
I'm having a really hard time saying "yes, let's integrate a random lib
with a wildly incompatible coding style, different calling ABI, a lib
which is in no distribution that I can see, and where the amount of
shimming I'd need to do would be comparable to the amount of lines it
would take for me to open-code the functionality I need". And the truth
is that libpmc does not offer a lot, because indeed there isn't a lot to
offer - basically definitions for the PTP management IDs and data sets,
a GETter and a SETter, a way to subscribe to push notifications, and an
abstraction over a socket, which works with L2/L4/UDS.

So this is the story of how I decided to implement PMC logic into
isochron "like a man". ptpmon (name was chosen differently because I
think "pmc" is very loaded these days) is a very spartan C file that is
more or less independent from the rest of the isochron project, and
which offers a very rudimentary API, basically just a GET for a
management data set. The ptpmon layer doesn't even get to care what that
management data set even is, it just takes a void * and the management
ID, provides the caller with a definition of the fields of that
structure, and Bob's your uncle.

I also decided to not support any other transport than UDS because I
don't need to. Compared to the checksync RFC where my attempt was to
patch ptp4l to accept remote subscriptions over Ethernet, here I am not
trying that. Instead, I'll just let the two instances of isochron
(sender and receiver) just connect individually to their local ptp4l
UNIX domain sockets, and then communicate between each other using their
stats socket.

That isn't supported right now - this patch merely adds the initial
ptpmon support, and its use in the isochron sender.

I don't use the subscription mechanism, and I am using the management
IDs defined by IEEE 1588 rather than the linuxptp non-portable ones
(TIME_STATUS_NP). Since I have no use case for anything other than
linuxptp, I might switch over to the non-portable TIME_STATUS_NP and
PORT_PROPERTIES_NP in the future, especially since the latter provides a
string with the port name, and I might correlate that with the
--interface argument of isochron rather than keep a list of all ports,
which I don't need. Right now I don't even know that I'm monitoring the
right clock of the right port.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>