isochron: create a management protocol between sender and receiver
Currently there is a simplistic management protocol that covers
statistics, and it is a bit hilarious: the receiver listens on a TCP
port for connections, the sender connects, and without saying anything,
the receiver opens the door, accepts that connection, dumps its log into
the socket and proceeds to close it.

With the introduction of ptpmon and sysmon, and the desire for the
sender to actually make sure that the receiver is in sync with it too,
we need to come up with some mechanism that is more elaborate, but also
extensible in case we need to make modifications in the future.

One other interesting issue is that the sender calls
prog_collect_rcv_stats() twice, and each time we open a TCP socket to
the isochron receiver leaving the other one unclosed and dangling.
This works because it is the receiver that closes it. Let's not do that,
because it just makes our work harder and impossible to extend.

After working on PTP management TLVs with ptpmon, that seems like a good
model, so we follow it here too.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>