isochron: forward ptpmon/sysmon info from receiver to sender
This uses the communication protocol created earlier to transfer sync
information queried by the isochron receiver all the way to the isochron
sender, in order for that to figure out whether now is a good time to
start test packet transmission or not.

The data is queried by the isochron receiver using its own UNIX domain
socket towards the ptp4l local to it.

The information transmitted between isochron instances is:

- phc2sys offset on receiver system (ISOCHRON_MID_SYSMON_OFFSET)
- ptp4l offset on receiver system (ISOCHRON_MID_PTPMON_OFFSET, must be
  processed along with ISOCHRON_MID_UTC_OFFSET to account for the
  UTC-to-TAI offset)
- port state to wait in case of transient state changes, or in case e.g.
  both ports are masters (ISOCHRON_MID_PORT_STATE)
- the clock identity of the grandmaster, in case the sender and receiver
  are not synchronized to the same thing (ISOCHRON_MID_GM_CLOCK_IDENTITY)

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>