isochron: look up the port state by interface name in ptpmon
Currently we use the MID_PORT_DATA_SET to query ptp4l port states, but
this is sub-optimal because we don't really know which port that is.

One case where that might not be ok is when ptp4l runs on an entirely
different port than the one on which we run isochron, and that port even
has a different PHC.

Unfortunately the standard PORT_DATA_SET doesn't give us something based
on which we can figure out whether this port is that port. IEEE 1588-2008
clause "7.5.2.2.2 IEEE EUI-64 clockIdentity values" doesn't say that the
clockIdentity should be derived in one way or another (ptp4l does seem
to derive it from the MAC address of the first port in the BC).

Enter the non-portable managementID PORT_PROPERTIES_NP which is defined
and implemented by ptp4l. This does provide the string corresponding to
the interface name, but is a bit hard to get. Its data type is a PTPText,
which is formatted Pascal-style, with the first octet denoting the
number of bytes in the string to follow.

I had expected that such a variable-length data type to only be present
at the end of a data set structure, on the premise that it would shift
all other following structure fields to unpredictable offsets, making it
impossible to map a variable-length C structure over the data set.

After opening the standard just now, I can see that the premise is
correct, but still that didn't stop the gods from putting PTPText fields
in the middle of data sets and structures, like:

| 5.3.10 FaultRecord
| The FaultRecord type is used to construct fault logs.
|
| struct FaultRecord
| {
| 	UInteger16 faultRecordLength;
| 	Timestamp faultTime;
| 	Enumeration8 severityCode;
| 	PTPText faultName;
| 	PTPText faultValue;
| 	PTPText faultDescription;
| };

Seriously?! Why not just make them fixed size if you're goint to do that?

Anyway, thanks to Jiri Benc for not doing that for PORT_PROPERTIES_NP in
commit 424bbde8fcca ("Custom management TLV PORT_PROPERTIES_NP").

Since I don't care about implementing a generic PTP client,
special-casing PTPText fields only at the end of structures seems to be
good enough, and a huge simplifying factor.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>