isochron: send: add initial sysmon support
sysmon is a more-or-less standalone translation module which deals with
supervising phc2sys. It does not communicate with phc2sys, instead it
receives the name of an Ethernet interface as argument, finds that
interface's PTP clock, and monitors that PTP clock's offset to
CLOCK_REALTIME (yes, we compare a UTC time base with a TAI time base,
and it is up to the caller of translating UTC to TAI).

Naturally, most of sysmon's code is shamelessly stolen from phc2sys.

Sadly, this new code is not free of warnings, and this is due to the
ethtool ioctl kernel headers:

/opt/gcc-arm-10.2-2020.11-x86_64-aarch64-none-linux-gnu/sysroot/usr/include/linux/ethtool.h:1575:35:
warning: comparison of integer expressions of different signedness: ‘__u32’ {aka ‘unsigned int’} and ‘int’ [-Wsign-compare]
 1575 |  return speed <= INT_MAX || speed == SPEED_UNKNOWN;
      |                                   ^~

So we'll have to live with it, it seems, just make Wsign-compare a
non-fatal warning.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>