isochron: require the sender to communicate the packet count it intends to send
I'm working with a board that appears to reorder my isochron data
packets (with hardware TX timestamping) relative to the statistics
packets (TCP, no timestamping). The test packets arrive later than the
sender's request for the log, and the sender thinks that packets are
missing, when in reality they haven't arrived by the time that the
receiver formulated the reply.

The solution is to be upfront and make the sender tell the receiver how
many packets are going to be sent. Then make the receiver wait for that
many packets to be received before actually forwarding the log.

This patch only deals with communicating the number of packets from the
sender to the receiver. There was some legacy behavior in the receiver,
but now that the sender always tells us what it's up to (and given the
fact that there have been so many changes incompatible with old binaries
recently), just drop the old behavior and now lazily allocate the log
buffer only when we know exactly how big it needs to be.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>